# dietgpu full float32 support