… dietgpu/float/GpuFloatCodec.h
… dietgpu/float/GpuFloatCompress.cu
… dietgpu/float/GpuFloatCompress.cuh
… dietgpu/float/GpuFloatDecompress.cuh
… dietgpu/float/GpuFloatUtils.cuh
… dietgpu/float_test.py