Summary:
This diff finishes adding full float32 compression support to dietgpu.

The 8 bit exponent is compressed via ANS, while the sign and significand bits are simply stored concatenated. At best we can compress to ~75% of original size. There is no zero compression implemented yet.

The format of the data to compress N float32 values deals with the fact that the uncompressed data is 3 bytes per float word, but there is no native 3 byte load/store:

```
- GpuFloatHeader (16 bytes)
- uint16 x N uncompressed data
  (N uint16 words containing the 2 low bytes of the 24 bit concatenated sign and significand)
- uint16 x (N - roundUp(N, 8)) padding
  (padding to ensure 16 byte alignment for subsequent fields)
- uint8 x N uncompressed data
  (N uint8 words containing the high 1 byte of the 24 bit concatenated sign and significand)
- uint8 x (N - roundUp(N, 16)) padding
  (padding to ensure 16 byte alignment for subsequent fields)
- GpuANSHeader (16 bytes)
<<< ANS compression data >>>
```

In order to implement this, I have temporary removed the vectorized ANS compressor and decompressor code (whereby each lane in a warp handles a contiguous block of 4 x uint8 words). This was possibly a premature optimization and will be revisited when I evaluate fused kernels.

Removing ANS vectorization hurts performance a bit:

**Before diff (baseline)**
```
Compressing [128 * 512 * 1024] torch.bfloat16
comp   time 0.504 ms B/W 266.4 GB/s, compression 134217728 B -> 90459584 B (0.6740x)
decomp time 0.403 ms B/W 333.4 GB/s
Compressing [128, [512 * 1024]] torch.bfloat16
comp   time 0.436 ms B/W 307.9 GB/s, compression 134217728 B -> 90435584 B (0.6738x)
decomp time 0.432 ms B/W 310.9 GB/s
Compressing [128 * 512 * 1024] torch.float16
comp   time 0.563 ms B/W 238.5 GB/s, compression 134217728 B -> 115520768 B (0.8607x)
decomp time 0.491 ms B/W 273.4 GB/s
Compressing [128, [512 * 1024]] torch.float16
comp   time 0.519 ms B/W 258.8 GB/s, compression 134217728 B -> 115548160 B (0.8609x)
decomp time 0.524 ms B/W 256.1 GB/s

```

**After diff (float32 + no ANS vectorization)**
```
Compressing [128 * 512 * 1024] torch.bfloat16
comp   time 0.481 ms B/W 279.0 GB/s, compression 134217728 B -> 90452656 B (0.6739x)
decomp time 0.511 ms B/W 262.6 GB/s
Compressing [128, [512 * 1024]] torch.bfloat16
comp   time 0.413 ms B/W 324.7 GB/s, compression 134217728 B -> 90476544 B (0.6741x)
decomp time 0.532 ms B/W 252.5 GB/s
Compressing [128 * 512 * 1024] torch.float16
comp   time 0.594 ms B/W 226.1 GB/s, compression 134217728 B -> 115516560 B (0.8607x)
decomp time 0.508 ms B/W 264.0 GB/s
Compressing [128, [512 * 1024]] torch.float16
comp   time 0.532 ms B/W 252.1 GB/s, compression 134217728 B -> 115558400 B (0.8610x)
decomp time 0.551 ms B/W 243.8 GB/s
Compressing [128 * 512 * 1024] torch.float32
comp   time 0.684 ms B/W 392.4 GB/s, compression 268435456 B -> 224653424 B (0.8369x)
decomp time 0.606 ms B/W 443.1 GB/s
Compressing [128, [512 * 1024]] torch.float32
comp   time 0.566 ms B/W 474.5 GB/s, compression 268435456 B -> 224647168 B (0.8369x)
decomp time 0.643 ms B/W 417.6 GB/s
```

Other changes:

- the batch prefix sum code no longer worked on batches larger than 512, seemingly due to our change from CUDA 11.1 to CUDA 11.4. I haven't investigated this in detail, but I think CUB register usage is going crazy (everything is being unrolled). This is now fixed by restricting batch prefix sum blocks to 512 max threads.

Differential Revision: D34811429